[#178] refactor: constants와 hooks 도메인 이동 (Phase 12)
- constants/level.ts → domains/place/constants/level.ts
- constants/key.ts 제거 (미사용)
- hooks/navigate.ts → domains/auth/hooks/useNavigateWithAuth.ts
- hooks/useImageDownload.ts → domains/jjikboul/hooks/
- hooks/useAppScheme.ts → domains/jjikboul/hooks/
- 모든 import 경로 수정
- constants 폴더 제거
- 빌드 테스트 통과
