+++ src/app/jjikboul/[jjikboulId]/page.tsx
@@ -5,8 +5,8 @@ import Image from "next/image";
 import { useJjikboul } from "@/domains/jjikboul/hooks/useJjikboul";
 import { useJjikboulUI } from "@/domains/jjikboul/hooks/useJjikboulUI";
 import { AsyncBoundary } from "@/lib/async";
-import useImageDownload from "@/hooks/useImageDownload";
-import useAppScheme from "@/hooks/useAppScheme";
+import useImageDownload from '@/domains/jjikboul/hooks/useImageDownload';
+import useAppScheme from '@/domains/jjikboul/hooks/useAppScheme';
 import Loading from "@/components/common/Loading";
 import Avatar from "@/components/common/Avatar";
 import { useGetJjikboulDetailQuery } from "@/domains/jjikboul/hooks/useGetJjikboulDetailQuery";
