--- src/constants/key.ts
@@ -1,1 +1,0 @@
-export const MEMBER_INFO_STORAGE_KEY = "memberInfo";
