+++ src/domains/auth/hooks/useNavigateWithAuth.ts
